Backend/requirements.txt
fastapi
uvicorn[standard]
requests
shodan
python-whois
psycopg2-binary
dnspython